Use Mergify for merge-when-green (#1107)
Merge when green does not support automatically updating the PR branch if it is out of date with master. When multiple people submit PRs at around the same time this can require multiple manual visits of the PR page to update against current master.

[Mergify](https://doc.mergify.io/index.html) seems to be a much more powerful tool that can also allow us to do things such as automatic rebases if the base branch gets merged, auto-merge dependabot changes etc.

This PR aims at just replacing the basic functionality we have today (merge when green) with mergify. This should also work on outdated branches.

### Test Plan
Wait for approval. Then push empty commit into master (to make it out of date) and attach merge-when-green label expect PR to be automatically merged after rebasing and running CI again.