Cancel transaction when solution submission time expires (#1089)
There is no reason to continue to try submitting the solution with
higher gas prices when the deadline is already over.
With this commit we cancel the transaction and exit when this happens.

Fixes #825 
Fixes #821 

### Test Plan
Added unit test for this. Wasn't able to test the full behavior with ganache but I did test that sending the noop transaction works in ganache.