Make pricegraph price source threaded
This is nice to have for the price estimator. It also makes sense here
as the calculation can be expensive. It does mean that its prices could
be slightly out of date compared to before this commit but considering
that we didn't have this price source at all previously this should be
fine.
While doing this I noticed a problem that we already had when all price
sources are threaded: technically we could start the solver without
having fetched prices at least once. The threaded price source updates
itself immediately but it is still possible this could take longer than
the code path that runs the solver in the first batch that we handle
after startup.