Handle tokens without price in rounding buffer calculation
The rounding buffer calculation takes token prices as input. If we don't
have a price for an order we need to pick some default or we can't
perform the calculation.
Previously we used `1` as the default but it seems more correct to reuse
the fee price (10e18) instead of 1. Either way is a guess but this is a
better heuristic.