Add rounding buffer to price estimator's orderbook
Some routes will make use of the pricegraph with the rounding buffer
applied and some without it. To facilitate this we store another
pricegraph instance and keep it up to date.
In a future PR `apply_rounding_buffer_to_auction_data` will be
implemented and routes will moved to the new pricegraph where
appropriate.