Increase gas price slower but higher (#1166)
We missed our 50k ETH/DAI auction today because the gas price was too high. This PR increases the cap to 200GWEI but makes it so that we only increase in 50% increments (doubling 90 GWEI on first attempt seems a bit harsh).

Going forward, we should not only implement #1099 but also make it that gas cap depends on the total fee that can be gained by the solver. In the example today we could have made $50 in fees. This should make us be willing to pay at least $50 in gas to get our tx mined.

### Test Plan
Unit tests