Make sure gas station gas price does not decrease (#1165)
The solution submission logic enforces that gas prices increase by at
least 12% each resubmission. However with the current logic it is
possible that the gas station returns significantly lower prices in the
next iteration which could cause this to be violated.
This commit ensures that we only ever increase the gas price estimate.

### Test Plan
Added new test for this.