Adjust gas price limit based on earned fees
With the current hardcoded maximum gas price limit we can run into
problems when gas prices are so high that the limit is not enough - even
when the submitting the solution would pay out more than the gas cost.
This commit changes the maximum gas prices to be a function of the gas
use of the solution, the fee payout and the cost of ether in the fee
(owl).