Price Estimator scaffolding for pricegraph with rounding buffer applied (#1142)
Some routes will make use of the pricegraph with the rounding buffer
applied and some without it. To facilitate this we store another
pricegraph instance and keep it up to date.
In a future PR `apply_rounding_buffer_to_auction_data` will be
implemented and routes will moved to the new pricegraph where
appropriate.

Also turned `fn get_pricegraph` into `fn pricegraph` because that is the norm in Rust and I had to touch all calls to this function anyway.


### Test Plan
The new pricegraph with rounding buffer applied isn't used anywhere yet. Tested that price estimator still works by running it and curling.