Invert priority price source (#1178)
Fixes #1150

We no longer override prices as a fallback. Quite to the contrary, on Saturday's token sale we were trying to use TokenData as an override to a faulty price estimation. This did however not work, because the hardcoded token data had less priority than the averaged price source.

### Test Plan
`export TOKEN_DATA='{ "T0072":{"alias":"MTA","decimals":18,"externalPrice":17} }'`

run it and see T0072 with price 17 in the instance file.