Fix race condition when cancelling solution submission transaction
When the solution submission transaction and the cancellation
transaction execute at the same time it is possible for us to observe
the failure first.
Previously we did not detect this case and could emit an error message
when there was not actually an error.
This commit handles this situation by waiting on and returning the
result of the other transaction if the first one completes with a "nonce
already used error".

This is part #1143 . In another PR a similar case in `retry.rs` will
be handled.