Make nonce already used error temporarily benign
This is a race condition that will be fixed in two future PRs. We always
had this race condition internally but it became visible in the logs
when introduced transaction cancelling when the submission takes too long.
Until those PRs had time to be tested in staging we want not be notified
of these errors.