Cache token infos at startup to prevent double fetch
When the cached token info fetcher does not yet contain a token it must
fetch it. If multiple calls `get_token_info` happen in parallel then we
likely make multiple requests to retrieve token info from the node
because of the way the code around the mutex works out.
This happens when the PriceOracle is created because we have multiple
price sources using the token info cache, and performing their initial
update at the same time.
This commit explicitly caches all tokens before the cache is passed to
the price sources.