Add enum for modeling amounts so that atoms are never fractional (#1202)
Related to #1111 

This PR defines a new `Amount` enum for representing token amounts in either atoms or base units. This has the effect of guaranteeing that atom amounts are never fractional. Note that the format is otherwise unchanged and base units are serialized as strings ("4.2" for example - see unit test) as there is still some discussion about how this should be done in the above mentioned issue.

Note that the transitive orderbook route does not use this type as it would be a "breaking" API change and we would need to evaluate with the FE team whether or not this can be changed. This should be easy to change in a follow up PR.

### Test Plan

Updated unit test and added new one. Additionally, run the price estimator and check that the results are in sync:
```
$ curl -s 'http://localhost:8080/api/v1/markets/7-1/estimated-buy-amount/1000000000000000000?atoms=true' | jq
{
  "baseTokenId": 7,
  "quoteTokenId": 1,
  "buyAmountInBase": "296069471770719551488",
  "sellAmountInQuote": "1000000000000000000"
}
$ curl -s 'http://localhost:8080/api/v1/markets/7-1/estimated-buy-amount/1?atoms=false' | jq                 
{
  "baseTokenId": 7,
  "quoteTokenId": 1,
  "buyAmountInBase": "296.06947177071953",
  "sellAmountInQuote": "1"
}
```