Fix race condition when retrying solution submission transaction (#1184)
Previously we would create the solution submission transactions with
increasing gas prices one after another and only check the result of the
most recent transaction. This leads to a race condition when a previous
transaction completes first.
This commit handles this situation by always checking all started
transactions and returning the first result that is not a "nonce already
used" error.

Fixes #1143

### Test Plan
Added unit tests for this situation.