[Easy] Use rinkeby gas staton on rinkeby (#1217)
The DIA team was trying to test their sale on rinkeby and is not getting any trades settled, because:

> Solution does not generate enough fees for the transaction to execute quickly enough: price cap 239628589 < fast gas price 93500000001

The gas price on rinkeby is usually around 1GWEI but we are using the mainnet estimates for determining economic viability. This PR makes it so that we use the appropriate URL.

### Test Plan
Run locally, see gas estimates for mainnet and rinkeby be different.