Update to stable Rust (#1214)
This PR updates the Rust version to the latest stable. The main motivation behind this are compiler and clippy lints being out of date causing build errors in #1211 and #1188. The issue here is that the lints are required by the nightly compiler (for the coverage tests for example) but don't exist in stable. This also requires the change in #1211 in order for the build to succeed. 

### Test Plan

The wonderful Rustc.