Make solver input/output files directory agnostic (#1226)
In order to make the driver run under the current binary setup (which is rooted under `/` but invoken from `/app`) and the mac compatible dockerized setup where we `cargo build` inside the container (rooted under /app/dex-services), we need to have the driver take working directory agnostic paths for the input and output directories.

This PR changes the paths from relative to absolute one.

Note the path for where we find the solver is still hardcoded to be under /app/batchauctions (or open-solver).

### Test Plan
Being able to run the solver from within the docker container (next PR). For now don't expect any difference, once merge check that AWS is still showing result files.