Use slightly higher min_avg_fee compared to max_gas_price
Currently we calculate the min avg fee and max gas price so that if the
solution earns exactly min avg fee it can be submitted at exactly fast
gas price while not losing money.
This is a problem when the gas price or eth price moves slightly between
the start of solution creation and submitting. To be more robust this
commit adds 10% to the min average fee.