[Trivial] Log verified solutions with correct stage (#1247)
While trying to pull the data from how many "verified" solutions we manage to submit I noticed that we don't have information on this in the graphana dashboard. We should be seeing a verified event in the processing times graph.

Reason for the absence is a copy paste error in the metrics code I introduced a few months back.

### Test Plan

Once merged I expect to see verified stage times in staging graphana.