Use new rounding buffer for estimated-amounts-at-price route (#1267)
We have previously discussed this in https://github.com/gnosis/dex-services/pull/1130/files#r454970573 .
I like this solution because there is no unbounded loop of expensive operations involved. It is correct in most cases and only 
has a small error otherwise. This small error (a change in the price the user wanted) is not a problem especially since I now notice that in the current version of the route we are already changing the price because we apply the rounding buffer to the sell amount.