Implement BaseUnits for Market route (#1315)
Second to last part of #1082

This PR implements base unit support for the markets route.

Thinking about how we can unify this type of logic, I'm wondering if it would make sense to introduce a trait `into_unit(unit, context)` that all our result types could implement (given the logic of how we convert atoms to base units likely depends on the operation). This would allow us moving logic away from the filters, but I'm not sure how much value there is in that.

### Test Plan
Added unit test + comparing the resulting book (atoms=false) with the visualization on Mesa. In particular picking some large orders and checking that volume and price correspond to the base unit values displayed on the web.