Introduce trait for mocking current time (#1319)
and show one example of how it would be used.

We've talked about this kind of trait before as an alternative to the "action pattern". What do you think of this implementation? I would like to use it in my refactor to get rid of the sleeping in stablex_driver and solution_submission.

### Test Plan
CI