Handle timing of solution submission in scheduler (#1322)
by splitting StableXDriver into two functions. This avoids sleeping in stablex_driver and is more appropriate in the scheduler.

PR is big because we get to delete a lot of test code. I wasn't able to split it up more and I already did not include handling earliest_solution_submit_time in the evm scheduler here.

### Test Plan
CI, new test