remove paginated orderbook reader (#1332)
This PR removes the paginated orderbook reader. This is because the on-chain filtered implementation is better in every way:
- Also EVM based
- Does not suffer from the same limitations as the paginated implementation

This way we no longer need to maintain an extra implementation that we arguably have no reason to use. Note that we also removed the `PaginatedAuctionDataReader` since it was no longer being used and clippy complained.

The question is, do you guys think we should also remove the:
1) Shadowed orderbook
2) Onchain filter orderbook

### Test Plan

Mostly removed code, adjusted unit tests where necessary.