introduce HTTP health reporting endpoint that runs in a backgroun thread (#1305)
This PR adds a new `HttpHealthEndpoint` component for that handles `/health/readiness` requests in the bacground HTTP server. It is currently only used in the price estimator (the driver always responds with "service unavailable").

This is in the spirit of providing a simple readiness endpoint so that we can do rolling updates in order to stay with 100% uptime even across restarts because of new deployments.

### Test Plan

Added unit tests. Start the price estimator:
```
$ cargo run -p price-estimator
```

While it is starting up, check that the `health` responds with a `503: Service Unavailable`:
```
$ curl -i http://localhost:9586/health/readiness
HTTP/1.1 503 Service Unavailable
Server: tiny-http (Rust)
Date: Tue, 18 Aug 2020 08:45:41 GMT
Content-Type: text/plain; charset=utf8
Content-Length: 19

service unavailable
```

Wait for the `price-estimator` to be ready:
```
2020-08-18T08:45:42.414Z INFO [price_estimator] Server ready.
```

Then check that the `health` endpoint responds correctly:
```
$ curl -i http://localhost:9586/health/readiness
HTTP/1.1 204 No Content
Server: tiny-http (Rust)
Date: Tue, 18 Aug 2020 08:45:43 GMT
Content-Length: 0
```