Add method for reading orderbook at block number (#1316)
This PR adds a new `StableXOrderbookReading::get_auction_data_for_block` trait method with additional implementations.

This would be an alternative to introducing a new trait for querying at a block number and timestamp in #1290. In a follow up PR I can add the method for querying at a specific timestamp as well.

### Test Plan

Added some new unit tests.