make /markets and /estimated-best-ask-price return consistent prices (#1348)
Fixes #1344

For more context see discussion in the issue. I added a long comment in the updated unit test justifying why the logic is now correct if the price is in quote.

Alternatively we could also just invert the price after it has been turned in the PriceEstimateResult type. Generally it would probably be nice to have some kind of meta information about what is base and what is quote in the result type (similar to how we have it for 

https://dex-price-estimator.gnosis.io/api/v1/markets/11-4/estimated-buy-amount/100000000?atoms=true

> {"baseTokenId":11,"quoteTokenId":4,"buyAmountInBase":"837137","sellAmountInQuote":"100000000"}

### Test Plan
Adjusted unit test