fetch batch for block number (#1356)
This PR modifies the `UpdatingOrderbook` implementation to read the timestamp for the block number so that the batch ID is more accurate.

### Test Plan

`EventRegistry` is tested by unit tests. Unfortunately the `UpdatingOrderbook` does not have unit tests and it does not seem trivial to set one up. Fortunately, only very little logic was added there (`updating_orderbook.rs`) so please review with extra attention. Suggestions for a better test plan are welcome!