Fix sleep not being awaited (#1372)
And added marked function as must_use. Usually rust warns about not awaiting the result of async functions but it looks like this doesn't trigger when returning BoxFuture. Technically with this reasoning we should put it everywhere where we are returning BoxFuture but not sure if that is really needed since usually we notice because we want to use the return value. sleep is special because we call it for the side effect.

### Test Plan
Tests don't notice this because mockall still registers the call.