Panic intead of endlessly allocating. (#1376)
This PR adds a new failing test similar to #1370 except triggered with an actual orderbook instead of one created directly on a `petgraph::Graph` as well as a `panic` to prevent the service from running out of memory in these cases. This will be addressed in #1370 and #1374 at which point the `#[should_panic]` can be removed.

This is just a band-aid to make it so the services don't restart in case we hit these rounding issues as a bandaid fix while the true fix is being worked on.

### Test Plan

Added assert and new unit test covering it.