Rename core crate to services-core
It turns out that it is bad idea to have a crate named `core`. The name
is already used in the standard library which can confuse for example
macros in dependencies that want to refer to the real `core` crate. I
encounted this when trying to use the async-trait macro
https://github.com/dtolnay/async-trait/issues/109 .