Convert TokenInfoFetching to async_trait
A welcome side effect of the conversion is that the mockall workaround
for lifetimes is no longer needed.
In this commit I also had to change one test slightly and move the the
default implementation of find_token_by_address into its own function.
Otherwise we cannot test it as mockall overrides the default
implementation.