Move StableXContract to async_trait
Here it did not work perfectly. Mockall only allows setting immediately
ready values on expectations. However for `submit_solution` and
`send_noop_transaction` we have tests that explicitly make use of
expectations that block for some time. This is not possible with
mockall. I filed a bug report.
Until that is fixed I've kept these methods as BoxFuture.