Change weight special value (#1395)
Running the fuzzer, I discovered an issue with the new `Weight`, specifically with the chosen special value for infinity. It turns out that the logic behind the range of values a `Weight` could have was a bit off, and the more natural choice of `i128::MAX` works better.

This PR updates the the `Weight::infiinite` special value and fixes the `Arbitrary` implementation to only provide `remaining_sell_amounts` less than, or equal the price denominator.

### Test Plan

Run the fuzzer and see it doesn't find a panic right away. Adjusted unit tests.