Allow disabling the dynamic economic viability calculation
We would like to be able to set fixed amounts for the min avg fee and
max gas price. This PR achieves that by making the subsidy factor
optional. If it is not set we only use default_min_avg_fee_per_order
and default_max_gas_price are used.