Allow setting fixed min-avg-fee-per-order (#1402)
We would like to be able to set fixed amounts for the min avg fee and
max gas price. This PR achieves that by making the subsidy factor
optional. If it is not set we only use default_min_avg_fee_per_order
and default_max_gas_price are used.

### Test Plan
I changed the code to print out the min avg fee to make sure it really does return the default. After merging I'd like to test this for one batch or so on staging.