Automatically determine chain id through web3
instead of having to specify it on the command line.

For the contract using the chain id instead of network id is what was
intended anyway and for the gas station it doesn't matter whether we
match on network id or chain id because they are the same for mainnet
and rinkeby.