Factor out duplicated econonomic viability struct creation (#1412)
It is better to not copy paste the code.

I got rid of the clap macro because we don't want to depend on that in services_core. Alternatively we could introduce a new crate for sharing functionality between the driver and the price estimator but I felt this is a good first step anyway and the new crate needs more discussion.

### Test Plan
No logic change, checked that both binaries still run.