[xDAI] Make EthPricing native token agnostic (#1419)
Part of 1418

EthPricing is really trying to estimate the price of the native token (ETH in the case of mainnet, xDAI in the case of DAI). This token is not always at ID 0.

This PR makes the token ID configurable and renames it to native_token to be compatible with non eth based side chains.

### Test Plan
Unit tests.