Filter orders with buy token == sell token (#1421)
In the pricegraph high level operations we already check that the two tokens in the market are different. I think the same makes sense for the orders in the orderbook as buying and selling the same token does not make sense.

Edit: This is already enforced in the smart contract so it probably makes more sense to enforce this when parsing the Element but we 'd have to add another possible error type to that function.

### Test Plan
Existing tests still pass and new test.