Use services_core http transport instead of web3 http transport
In preparation for the ethcontract upgrade. In the new ethcontract
version we get a new web3 version whose default http transport uses
hyper (tokio) which breaks the code here that uses `futures::block_on`
because these futures can only be executed on the tokio runtime.
We avoid this by using the custom http transport we already have for the
driver. Its futures can be run on any executor.
Another benefit of this change that is that the hyper and tokio don't
need to be compiled for the e2e test.