Add mix of dynamic and static economic viability strategy
called DynamicBoundedByStatic: Use Dynamic first. If it fails or the
result is worse (larger min-avg-fee, lower max-gas-price) than Static,
use Static instead.

This is useful because for ease of use we want to guarantee a minimum
viable order size but still lower it when this is economically viable
because of low gas price.