Give more control over economic viability computation
Instead of storing an EconomicViabilityComputing in
OptimizationPriceFinder we pass the min_avg_fee directly in the
find_prices method.
This makes the price finding code simpler because it has to access one
les external interface and gives us more control over the economic
viability strategy in StableXDriver.
I plan on making use of to implement a new economic viability strategy
which is a mix of dynamic and static.