Do not use fallback min_avg_fee in Dynamic strategy
Initially we only had the dynamic strategy and used fallback min_avg_fee
and max_gas_price in case getting the gas price or ether price would
fail.
Now these options are also used for the static strategy.
I want to remove them from the dynamic strategy because it is misleading
to use the same config value for different things and we rely on a
working gas station and token price estimation anyway. It would be
unexpected that we fallback to very different fallback values when the
strategy is Dynamic. It is better to have this be a real error.