Add more fine grained gas price estimation method (#1493)
In preparation for using more advanced gas estimation this commit adds a
new method to the trait which estimates gas price based on the
transaction's gas use and time limit.

After this is merged I want to implement the trait for other "gas stations" than the gnosis one.

### Test Plan
Ci. New function is not used anywhere.