Bounded bellman ford (#1495)
Part of #917. This PR is basically a copy of of #1368 (all credit to @fedgiac who did all the hard work) but implementing the unbounded case in its own submodule. More information on the logic can be found in the original PR.

I did a pass over the comments from back then and tried to apply them (still missing a few unit tests).

### Test Plan

The unit tests @fedgiac added still pass.