Update default values for Economic Viability config (#1513)
We are facing an issue of overlapping docker-compose file used both in CI testing and also when following the REDME for running a solver in a container (specifically for macOS). This PR is intended to update those default values to something less dangerous and also keep some undesired overlap between CI configurations and those used in practice.

Fixes #1512


### Test Plan
CLI passes without any logical changes.