Reduce full orderbook sync memory usage (#1511)
By making use of a change in ethcontract that creates a stream for past
events instead of one vector with all of them.
We chunk the stream again so that getting the block timestamps stays
fast as we use a batch transport for that.

This switches to  a specific commit of ethcontract so that we get the above mentioned change. Once ethcontract makes a release with this change we can switch back to a cargo version.

### Test Plan
Unfortunately UpdatingOrderbook cannot easily be tested in the current form because we don't have a mockable web3 trait and because it does file io. I would like to change that but not sure what the best approach is yet and this wouldn't be part of this PR.