Reduce gas estimator error logs (#1525)
The gas now estimator from time to time fails single requests for which we get an alert because this is an error log. Since we use multiple estimators if previous one fails this is not a problem and I don't see value in having this be an alert.

The easiest way to solve this would be to turn the error log into a warning, which I find acceptable too.
This PR instead starts off logging warnings but after an estimator failed N times in a row starts logging an error.

This is has the advantage that we would notice an estimator going down permanently or changing their api in a backwards incompatible way. On the other hand it could still be annoying to get an alert when an estimator goes down for 10 minutes when this hasn't actually impacted our service because of the fallback estimators. I am not sure what others prefer so let's discuss that in this PR.

### Test Plan
Temporarily add a test that with always failing estimator and observe it's errors (also change log to println).