[GasEstimate] refactor gas estimators to use abstract transport trait (#1522)
This will allow extracting the gas estimation logic into its own crate and reuse it e.g. in the gas token miner as well as in oba services without requiring logic that is specific to `services-core` (e.g. HTTPTransport, MetricLabels, etc). 
The `eth_node` implementation will likely not be extractable as it is bound to the web3 struct and doesn't add much custom logic.

### Test Plan
CI + `cargo test gas_price -- --ignored`