More aggressive gas price increase
It can happen that the new gas price is only slightly larger than the
old one, without being large enough to respect the minimum gas price
increase.
The old algorithm would in this case not resubmit the solution.
The new algorithm always increases the gas price to the minimum increase
if the new gas price is larger than the old gas price.