Simplify trait bounds fro AsyncSleeping, Now, SolutionTransactionSending
For AsyncSleeping and and Now we can safely assume that that this trait
is always 'static which gets rid of some generic lifetimes.
For SolutionTransactionSending we can assume that it is always Send and
Sync but here I didn't want to add 'static because it would be
reasonable to create an implementation based on &StablexContract instead
of Arc for example.