Make the retry api agnostic of how transactions are sent
which is is a big step towards making this a standalone crate. We
accomplish this by using traits to encapsulate how the "solution" and
the cancellation is sent. This way there is no dependency on
StablexContract or even web3.
solution_submission has intentionally not yet been fully updated to the
new api to keep the PR smaller.