Use SolutionSubmission Wrapper contract (#1544)
With this PR we will be able to use the GasToken to submit solutions on mainnet. This is achieved by funneling solutions through a WrapperContract, which has the same interface as the original BatchExchange but on top of forwarding it to the original one it also burns a bunch of gas tokens (if the current gas price is above a certain threshold). Currently the threshold is set to 60 GWEI in the smart contract

### Test Plan
Unit tests still pass. Since this requires a solution submission to test, my plan is to deploy it to staging and check execution of the first couple of solutions.