Make the retry api agnostic of how transactions are sent (#1548)
which is is a big step towards making this a standalone crate. We
accomplish this by using traits to encapsulate how the "solution" and
the cancellation is sent. This way there is no dependency on
StablexContract or even web3.
solution_submission has intentionally not yet been fully updated to the
new api to keep the PR smaller.

### Test Plan
Tests for retry pass but tests for solution_submission have been commented out and code contains a `todo!` so CI won't pass.