[Easy] Fix inverted logic in was_mined() (#1558)
Fixes alerts we have seen twice today in dev channel.

If the result of a Cancellation/Solution submission is not an err, it was mined. At the moment it returns false in that case leading to successful txs not being recognized up by the system.

### Test Plan

Will add a unit test.