Use transaction retry and gas estimation dependencies from git (#1626)
because we want to also use them in gp-v2.

https://github.com/gnosis/gp-gas-estimation
https://github.com/gnosis/gp-transaction-retry

I want to use them as git dependencies from gp v1 and v2 first.  We could eventually publish them on crates.io with the aim of making it easy for projects outside of gp to use them too. That is more work because we should polish them more so I do not want to commit to that now.
The current version of the crates is exactly what was in this repository before. Before using them in v2 we need to upgrade some dependencies which I am going to do as PRs to those repositories. After that v2 can depend on them.

### Test Plan
CI